+++ src/app.module.ts
@@ -1,6 +1,7 @@
 import { MikroOrmModule } from "@mikro-orm/nestjs";
 import { Module } from "@nestjs/common";
 import { ConfigModule } from "@nestjs/config";
+import { EventEmitterModule } from "@nestjs/event-emitter";
 import { CrashGamesModule } from "~modules/crash-games/crash-games.module";
 import { UserTemplatesModule } from "~modules/user-templates/user-templates.module";
 import { UsersModule } from "~modules/users/users.module";
@@ -12,6 +13,7 @@ import { UsersModule } from "~modules/users/users.module";
       isGlobal: true
     }),
     MikroOrmModule.forRoot(),
+    EventEmitterModule.forRoot(),
     CrashGamesModule,
     UserTemplatesModule,
     UsersModule