+++ src/common/enums/ws-events.enum.ts
@@ -1,9 +1,13 @@
 export enum WebsocketEventsEnum {
-  CG_DATA = "data",
+  CG_DATA = "crash-game/data",
 
-  CG_ADD_BET = "add-bet",
-  CG_ADD_BET_REPLY = "add-bet-reply",
+  CG_ADD_BET = "crash-game/add-bet",
+  CG_ADD_BET_REPLY = "crash-game/add-bet-reply",
 
-  CG_CASHOUT = "cashout",
-  CG_CASHOUT_REPLY = "cashout-reply"
+  CG_CASHOUT = "crash-game/cashout",
+  CG_CASHOUT_REPLY = "crash-game/cashout-reply",
+
+  CG_EM_CREATE = "crash-game.create",
+  CG_EM_START = "crash-game.start",
+  CG_EM_END = "crash-game.end"
 }